Switch to client/server stimulus generation
Stimulus is now provided by a seperate client using ZeroMQ to
communicate with the server which uses cocotb to run the simulation.
This model allows the client to run on a different python environment
and provides a more flexible 'push' system where the simultation steps
when given new stimulus as opposed to a 'pull' system where the
simulation requests new stimulus within the same python process.